+++ owlcms/src/main/frontend/components/Results.js
@@ -17,7 +17,7 @@ class Results extends LitElement {
       <link rel="stylesheet" type="text/css" .href="${"local/" + (this.stylesDir ?? "") + "/results" + (this.autoversion ?? "") + ".css"}" />
       <link rel="stylesheet" type="text/css" .href="${"local/" + (this.stylesDir ?? "") + "/resultsCustomization" + (this.autoversion ?? "") + ".css"}" />
 
-      <div class="${this.wrapperClasses()}" style="${this.sizeOverride}">
+      <div class="${this.wrapperClasses()}" style="${this.sizeOverride} ${this.colorOverride}">
         <div class="blockPositioningWrapper">
           <div class="waiting" style="${this.waitingStyles()}">
             <div>
@@ -299,6 +299,7 @@ class Results extends LitElement {
       teamWidthClass: {},
       sizeOverride: {},
       twOverride: {},
+	    colorOverride: {},
       video: {},
       showLiftRanks: {type: Boolean},
       showBest: {type: Boolean},
@@ -314,6 +315,7 @@ class Results extends LitElement {
       javaComponentId: {},
       stylesDir: {},
       autoVersion: {},
+
     };
   }